+++ src/core/calendar/Update.ts
@@ -1,43 +1,17 @@
 import { match } from "../../helpers/match";
-import type { ActionType, Model, SubmitData } from "../../types";
+import type { ActionType, Model } from "../../types";
 import { MSGS } from "../../types";
-import { createCommandManager } from "../command";
 import { updateActiveDay } from "./update/updateActiveDay";
 import { updateCurrentMonth } from "./update/updateCurrentMonth";
+import { updateCurrentSlotId } from "./update/updateCurrentSlotId";
 import { updateGoToToday } from "./update/updateGoToToday";
 import { updateNextMonth } from "./update/updateNextMonth";
+import { updateOnSubmit } from "./update/updateOnSubmit";
 import { updatePrevMonth } from "./update/updatePrevMonth";
 import { updateSelectedDate } from "./update/updateSelectedDate";
+import { updateSetEventsBeforeAddingSlot } from "./update/updateSetEventsBeforeAddingSlot";
 import { updateShowAddForm } from "./update/updateShowAddForm";
-
-export function onSubmitMsg(submitData: SubmitData) {
-  return {
-    type: MSGS.ON_SUBMIT,
-    submitData,
-  };
-}
-
-export function currentSlotIdMsg(slotId: number | null) {
-  return {
-    type: MSGS.CURRENT_SLOT_ID,
-    slotId,
-  };
-}
-
-export function setEventsBeforeAddingSlotMsg(
-  eventsBeforeAddedSlot: Model["events"]
-) {
-  return {
-    type: MSGS.SET_EVENTS_BEFORE_ADDING_SLOT,
-    eventsBeforeAddedSlot,
-  };
-}
-
-export function undoAddLatestSlotMsg() {
-  return {
-    type: MSGS.UNDO_ADD_LATEST_SLOT,
-  };
-}
+import { updateUndoLatestSlot } from "./update/updateUndoLatestSlot";
 
 export default function update(msg: ActionType, model: Model): Model {
   return match<ActionType, Model>(msg)
@@ -71,51 +45,19 @@ export default function update(msg: ActionType, model: Model): Model {
     )
     .on(
       (x) => x.type === MSGS.CURRENT_SLOT_ID,
-      (x) => {
-        if (x.type !== "CURRENT_SLOT_ID") return model;
-
-        return {
-          ...model,
-          currentSlotId: x.slotId,
-        };
-      }
+      (x) => updateCurrentSlotId({ msg: x, model })
     )
     .on(
       (x) => x.type === MSGS.SET_EVENTS_BEFORE_ADDING_SLOT,
-      (x) => {
-        if (x.type !== "SET_EVENTS_BEFORE_ADDING_SLOT") return model;
-
-        return {
-          ...model,
-          eventsBeforeAddedSlot: x.eventsBeforeAddedSlot,
-        };
-      }
+      (x) => updateSetEventsBeforeAddingSlot({ msg: x, model })
     )
     .on(
       (x) => x.type === MSGS.UNDO_ADD_LATEST_SLOT,
-      () => {
-        const undoModel = { ...model, events: model.eventsBeforeAddedSlot };
-        const undoManager = createCommandManager(undoModel, msg);
-
-        undoManager.undo();
-
-        return {
-          ...undoModel,
-        };
-      }
+      (x) => updateUndoLatestSlot({ msg: x, model })
     )
     .on(
       (x) => x.type === MSGS.ON_SUBMIT,
-      () => {
-        const newModel = { ...model };
-        const submitManager = createCommandManager(newModel, msg);
-
-        submitManager.doCommand("ADD_SLOT");
-
-        return {
-          ...newModel,
-        };
-      }
+      (x) => updateOnSubmit({ msg: x, model })
     )
     .otherwise(() => model);
 }
