+++ src/core/calendar/update/updateCurrentSlotId.ts
@@ -1,0 +1,23 @@
+import { ActionType, Model, MSGS } from "../../../types";
+
+export function currentSlotIdMsg(slotId: number | null) {
+  return {
+    type: MSGS.CURRENT_SLOT_ID,
+    slotId,
+  };
+}
+
+export const updateCurrentSlotId = ({
+  msg,
+  model,
+}: {
+  msg: ActionType;
+  model: Model;
+}) => {
+  if (msg.type !== MSGS.CURRENT_SLOT_ID) return model;
+
+  return {
+    ...model,
+    currentSlotId: msg.slotId,
+  };
+};
